requirements.txt
selenium==4.19.0
pytest==8.1.1
pytest-html==4.1.1
pytest-xdist==3.5.0
pytest-rerunfailures==13.0
python-dotenv==1.0.1